Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
supabase==2.3.4
python-dotenv==1.0.0
jwt==1.3.1
requests==2.31.0